api: get cloudflare worker using cli params, handling requests
