+++ packages/api/src/worker-secrets.json
@@ -1,1 +1,6 @@
-{}
+{
+  "LP_STORAGE": "cloudflare",
+  "LP_CLOUDFLARE_NAMESPACE": "efceebc4f0d947e2b69df96c526a6526",
+  "LP_CLOUDFLARE_ACCOUNT": "2d8e3e6f3567c9760d12a19ca2b931b3",
+  "LP_CLOUDFLARE_AUTH": "9hXDiBeZ9aGeWhwxQRsrtMpnrwwNtpszE6Wgvnw0"
+}
